Move rootDomain
