+++ modules/enrichmentFpdModule.js
@@ -1,74 +1,20 @@
-
 /**
  * This module sets default values and validates ortb2 first part data
  * @module modules/firstPartyData
  */
-import { timestamp, mergeDeep } from '../src/utils.js';
-import { submodule } from '../src/hook.js';
+import {mergeDeep} from '../src/utils.js';
+import {submodule} from '../src/hook.js';
 import {getRefererInfo, parseDomain} from '../src/refererDetection.js';
-import { getCoreStorageManager } from '../src/storageManager.js';
 import {GreedyPromise} from '../src/utils/promise.js';
 import {getHighEntropySUA, getLowEntropySUA} from '../libraries/fpd/sua.js';
+import {findRootDomain} from '../src/fpd/rootDomain.js';
+
+export {coreStorage} from '../src/fpd/rootDomain.js';
 
 let ortb2;
 let win = (window === window.top) ? window : window.top;
-export const coreStorage = getCoreStorageManager('enrichmentFpd');
 
 export const sua = {he: getHighEntropySUA, le: getLowEntropySUA};
-
-/**
-  * Find the root domain
-  * @param {string|undefined} fullDomain
-  * @return {string}
-*/
-export function findRootDomain(fullDomain = window.location.hostname) {
-  if (!coreStorage.cookiesAreEnabled()) {
-    return fullDomain;
-  }
-
-  const domainParts = fullDomain.split('.');
-  if (domainParts.length == 2) {
-    return fullDomain;
-  }
-  let rootDomain;
-  let continueSearching;
-  let startIndex = -2;
-  const TEST_COOKIE_NAME = `_rdc${Date.now()}`;
-  const TEST_COOKIE_VALUE = 'writeable';
-  do {
-    rootDomain = domainParts.slice(startIndex).join('.');
-    let expirationDate = new Date(timestamp() + 10 * 1000).toUTCString();
-
-    // Write a test cookie
-    coreStorage.setCookie(
-      TEST_COOKIE_NAME,
-      TEST_COOKIE_VALUE,
-      expirationDate,
-      'Lax',
-      rootDomain,
-      undefined
-    );
-
-    // See if the write was successful
-    const value = coreStorage.getCookie(TEST_COOKIE_NAME, undefined);
-    if (value === TEST_COOKIE_VALUE) {
-      continueSearching = false;
-      // Delete our test cookie
-      coreStorage.setCookie(
-        TEST_COOKIE_NAME,
-        '',
-        'Thu, 01 Jan 1970 00:00:01 GMT',
-        undefined,
-        rootDomain,
-        undefined
-      );
-    } else {
-      startIndex += -1;
-      continueSearching = Math.abs(startIndex) <= domainParts.length;
-    }
-  } while (continueSearching);
-  return rootDomain;
-}
 
 /**
  * Checks for referer and if exists merges into ortb2 global data
