+++ src/fpd/enrichment.js
@@ -1,0 +1,5 @@
+import {hook} from '../hook.js';
+
+export const enrichFPD = hook('sync', (orbt2) => {
+
+});
